Replace deprecated properties
This fixes the following theme parsing errors using GTK 3.24:
```
Gtk-WARNING **: 19:33:09.442: Theme parsing error: gtk-widgets.css:16:32: The style property GtkExpander:expander-size is deprecated and shouldn't be used anymore. It will be removed in a future version

Gtk-WARNING **: 19:33:09.442: Theme parsing error: gtk-widgets.css:17:35: The style property GtkExpander:expander-spacing is deprecated and shouldn't be used anymore. It will be removed in a future version

Gtk-WARNING **: 19:33:09.442: Theme parsing error: gtk-widgets.css:23:47: The style property GtkProgressBar:min-horizontal-bar-height is deprecated and shouldn't be used anymore. It will be removed in a future version

Gtk-WARNING **: 19:33:09.442: Theme parsing error: gtk-widgets.css:24:44: The style property GtkProgressBar:min-vertical-bar-width is deprecated and shouldn't be used anymore. It will be removed in a future version
```